pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.39",
    "flask-wtf>=1.2.2",
    "flask-pymongo>=3.0.1",
    "flask-bcrypt>=1.0.1",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "wtforms>=3.2.1",
]
